refactor!: move PublicTopic to Zoe contractSupport
