+++ packages/notifier/src/index.js
@@ -20,4 +20,3 @@ export {
 } from './asyncIterableAdaptor.js';
 export * from './storesub.js';
 export * from './stored-notifier.js';
-export * from './stored-topic.js';
